requirements.txt
pandas==2.2.3
pytest==8.3.4
python-dotenv==1.0.1
Requests==2.32.3
